Tests/LibWeb/Text/expected/document-caret-position-from-point-shadow-roots.txt

without shadowRoots retargets to host parent: true
without shadowRoots retargets to host index: true
unrelated shadowRoot still retargets to host parent: true
unrelated shadowRoot still retargets to host index: true
allowed shadowRoot exposes text node: true
allowed shadowRoot exposes text offset: true
